Raise `ArgumentError` instead of a client error.
The reasoning behind this is that errors inheriting from
`GOVUK::Client::Errors::BaseError` represent something going wrong in
the API itself (be that a validation error, server error or network
problem). For cases of programmer error (or, if you prefer, violations
of a method's preconditions), we raise the appropriate standard Ruby
error.